Moved cli helper functions to utils module. Also, clarified unauthenticated gh message.
