+++ src/gh_profiler/cli.py
@@ -7,6 +7,7 @@ import click
 
 from . import gh_profiler
 from .utils.infra_utils import run_cmd
+from .utils import cli_utils
 
 
 @click.command()
@@ -37,52 +38,5 @@ def main(target):
     
     # The user provided a PR/issue number. Get the relevant username, then
     # call gh_profiler.main().
-    username = _get_username(pr_issue_num)
+    username = cli_utils.get_username(pr_issue_num)
     gh_profiler.main(username)
-
-
-# --- Helper functions ---
-
-def _get_username(pr_issue_num):
-    """Get the user that opened this PR/issue."""
-    repo_slug = _get_repo_slug()
-    
-    # Try as a PR, then as an issue.
-    if username := _process_pr(pr_issue_num, repo_slug):
-        return username
-    if username := _process_issue(pr_issue_num, repo_slug):
-        return username
-
-    msg = f"Couldn't find a PR or issue #{pr_issue_num} in the repository {repo_slug}."
-    sys.exit(msg)
-
-def _get_repo_slug():
-    """Ask `gh` for the resolved default repo (honors `gh repo set-default`)."""
-    slug = run_cmd("gh repo view --json nameWithOwner --jq .nameWithOwner").strip()
-    if not slug:
-        msg = (
-            "Couldn't determine the default GitHub repository. "
-            "Run `gh repo set-default` in this directory and try again."
-        )
-        sys.exit(msg)
-    return slug
-
-def _process_pr(pr_issue_num, repo_slug):
-    """See if this is a PR."""
-    pr_cmd = f'gh pr view {pr_issue_num} --repo {repo_slug} --json author --jq ".author.login"'
-    try:
-        if username := run_cmd(pr_cmd).strip():
-            return username
-    except Exception as e:
-        breakpoint()
-        return None
-
-def _process_issue(pr_issue_num, repo_slug):
-    """See if this is an issue."""
-    issue_cmd = f'gh issue view {pr_issue_num} --repo {repo_slug} --json author --jq ".author.login"'
-    try:
-        if username := run_cmd(issue_cmd).strip():
-            return username
-    except Exception:
-        msg = f"Couldn't find a PR or issue #{pr_issue_num} in the repository {repo_slug}."
-        sys.exit(msg)
